compose: search: Force single pane layout for app details while searching
This looks awful on foldables as it still draws based on the window width and not the
pane width. I cannot find a simple way to differentiate the width between foldables and tablet
to display the side pane here so just disable it for now.

Hopefully, Google will add a way to adapt the adaptive layouts based on pane width.

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>